requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart
Flask==2.3.3
Flask-CORS==4.0.0
PyJWT==2.8.0